Create Random math quiz generator.py
Write a program that gives simple math quizzes. The program should display two random
numbers that are to be added, such as:
247
+ 129
The program should allow the student to enter the answer. If the answer is correct, a message of
congratulations should be displayed. If the answer is incorrect, a message showing the correct
answer should be displayed.