Create Total Pay Calculation.py
Write a Python function to compute the total pay given to an employee, which is a product of the
hourly rate times the hours worked. Do note that the employee will be paid 1.5 times the hourly
rate for hours worked above 40 hours (overtime). Create a function called computepay which
takes two parameters (hours and rate).